+++ docs/fr_FR/changelog.md
@@ -4,6 +4,15 @@
 >
 >S'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+# 23/09/2024 Beta 1.0.4
+- Option pour exclure ou non les datas du backup Jeedom
+- Ajout de la pause PTZ
+- Ajout commande statut serveur et disponibilité serveur
+- Save automatiquement le bbox sur les snapshots
+- Optimisation du cron
+- Correction erreur file_get_content si fichiers n'existe pas
+- Correction filtre date (Firefox)
+
 # 21/09/2024 Beta 1.0.3
 - Option pour flux RTSP (voir documentation)
 - Force le type génèrique de l'url snapshot (faites une recherche ou un save de chaque equipement)